+++ CHANGELOG.md
@@ -51,6 +51,8 @@ Changelog for `ta4j`, roughly following [keepachangelog.com](http://keepachangel
 - Added `SMMAIndicator` to `indicators.averages`-package: Smoothed Moving Average (SMMA) indicator
 - Added `SGMAIndicator` to `indicators.averages`-package: Savitzky-Golay Moving Average (SGMA) indicator
 - Added `LSMAIndicator` to `indicators.averages`-package: Least Squares Moving Average (LSMA) indicator
+- Added `KiJunV2Indicator` to `indicators.averages`-package: Kihon Moving Average (KiJunV2) indicator
+
 
 ## 0.17 (released September 9, 2024)